[rename] myActivitiesApi.js 대문자 시작 => 소문자로 변경
